+++ src/column-layout/index.tsx
@@ -1,0 +1,23 @@
+import cn from 'classnames';
+import { CSSProperties, ReactNode } from 'react';
+import './index.css';
+
+export interface ColumnLayoutProps {
+  align?: 'center' | 'flex-start' | 'flex-end' | 'stretch';
+  children?: ReactNode;
+  className?: string;
+  gap?: number | string;
+  style?: CSSProperties;
+  wrap?: boolean;
+}
+
+export function ColumnLayout({ align, children, className, gap, style, wrap }: ColumnLayoutProps) {
+  return (
+    <div
+      className={cn('gyh-column-layout', wrap && 'gyh-column-layout-wrap', className)}
+      style={{ ...style, gap, alignItems: align }}
+    >
+      {children}
+    </div>
+  );
+}