+++ src/index.tsx
@@ -1,1 +1,2 @@
+export * from './column-layout';
 export * from './row-layout';