+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Unreleased
 
+- ColumnLayout: new component
 - RowLayout: new component